test: MockMemoryManager moves to tests._utils; private managers built directly
The type(low_memory)(...) indirection and its low_memory fixture
dependency are gone: tests that need a private mock manager import
MockMemoryManager from tests._utils and construct it with the limit
they mean, without touching the shared session manager.
